// client/src/main/cv-viewer/src/app/api/model/user.ts
/**
 * CV Viewer
 * CV Viewer.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: neilje123@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface User { 
    userName?: string;
    userPassword?: string;
    userRoles?: Array<string>;
}